Updating organization of APOD SideBar
