+++ src/pages/APOD/components/SecondPage/Tabs/SecondTab/styles.ts
@@ -1,25 +1,55 @@
 import styled from "styled-components";
 import { RightArrowAlt } from "@styled-icons/boxicons-regular/RightArrowAlt";
-
 export const Container = styled.div`
+  margin-top: 5%;
+  width: 100%;
+  overflow-x: auto;
+`;
+export const Wrapper = styled.div`
+  width: 100%;
+`;
+export const MyNavBar = styled.div`
+  width: 100%;
   display: flex;
-  flex-direction: column;
-  max-width: calc(100% - 300px);
-  margin-left: 25%;
-  @media (max-width: 1100px) {
-    margin: auto;
-    max-width: 95%;
+  justify-content: flex-end;
+`;
+
+export const MyLink = styled.div`
+  margin-right: 20px;
+  @media (max-width: 1200px) {
+    margin-bottom: 15px;
   }
-  max-height: 100vh;
-  overflow-y: auto;
-  ::-webkit-scrollbar {
-    display: none;
+  justify-content: center;
+  display: flex;
+  > span {
+    cursor: pointer;
+    width: max-content;
+    font-family: "Poppins";
+    font-style: normal;
+    font-weight: normal;
+    font-size: 18px;
+    line-height: 27px;
+    color: black !important;
+
+    position: relative;
+    &.ActiveSpanSecondPage {
+      ::after {
+        position: absolute;
+        width: calc(100% - 20px);
+        left: 10px;
+        bottom: -9px;
+        content: "";
+        height: 5px;
+        border-radius: 5px;
+        background-color: var(--background);
+      }
+    }
   }
 `;
 
 export const FormContent = styled.form`
   padding: 10px;
-  margin: 10% 0 3% 3%;
+  margin: 0 0 3% 3%;
   @media (max-width: 600px) {
     margin: 40% 0 5% 3%;
   }
@@ -41,9 +71,6 @@ export const Title = styled.div`
   font-style: normal;
   font-weight: 600;
   font-size: 24px;
-  @media (max-width: 768px) {
-    font-size: 18px;
-  }
   line-height: 36px;
   color: white;
 `;
@@ -55,7 +82,13 @@ export const MyInput = styled.input`
   border-radius: 8px;
   background-color: var(--secundary);
 `;
-
+export const ImageContent = styled.div`
+  background-color: rgba(18, 32, 47, 1);
+  border-radius: 8px;
+  text-align: center;
+  max-width: 1100px;
+  margin: 3% 20px;
+`;
 export const ArrowIcon = styled(RightArrowAlt)`
   width: 40px;
   height: 40px;
@@ -64,9 +97,25 @@ export const ArrowIcon = styled(RightArrowAlt)`
   cursor: pointer;
   flex-shrink: 0;
 `;
+
+export const MyImage = styled.img`
+  max-width: 1000px;
+  border-radius: 5px;
+  @media (max-width: 1500px) {
+    margin-top: 0;
+    width: 70%;
+    margin-left: 0;
+    padding: 20px;
+    margin: auto;
+  }
+  @media (max-width: 1100px) {
+    width: 100%;
+    padding: 0;
+  }
+`;
 export const TextContent = styled.div`
   max-width: 1100px;
-  padding: 10px;
+  padding-bottom: 10px;
 `;
 export const TitleCarousel = styled.div`
   font-family: "Poppins";
@@ -85,40 +134,13 @@ export const Text = styled.div`
   font-family: "Poppins";
   font-style: normal;
   font-weight: normal;
-  font-size: 20px;
-  @media (max-width: 768px) {
-    font-size: 13px;
-    line-height: inherit;
-  }
+  font-size: 16px;
   line-height: 36px;
   color: white;
   text-align: justify;
-  padding-right: 5px;
-  @media (max-width: 1500px) {
-    font-size: 15px;
-  }
-`;
-export const ImageContent = styled.div`
-  background-color: rgba(18, 32, 47, 1);
-  border-radius: 8px;
-  padding: 12px;
-  text-align: center;
-  max-width: 1100px;
-  margin: 3% 20px;
-`;
-export const MyImage = styled.img`
-  max-width: 1000px;
-  border-radius: 5px;
-  @media (max-width: 1500px) {
-    margin-top: 0;
-    width: 70%;
-    margin-left: 0;
-    padding: 20px;
-    margin: auto;
-  }
-  @media (max-width: 1100px) {
-    width: 100%;
-    padding: 0;
+  @media (max-width: 768px) {
+    font-size: 13px;
+    line-height: inherit;
   }
 `;
 export const VideoContent = styled.div`
